Moved DMG-related functions into new UtilityMethods Swift file (Issue #53)
